+++ CHANGELOG.md
@@ -23,6 +23,10 @@ _None_
   [Ilya Puchka](https://github.com/ilyapuchka)
   [#254](https://github.com/stencilproject/Stencil/pull/254)
 
+- Throw syntax error on empty variable tags (`{{ }}`) instead `fatalError`.  
+  [Ilya Puchka](https://github.com/ilyapuchka)
+  [#263](https://github.com/stencilproject/Stencil/pull/263)
+
 ### Internal Changes
 
 - `Token` type converted to struct to allow computing token components only once.